refine msrr and integrate
