+++ Cargo.toml
@@ -28,6 +28,9 @@ libc = "0.2.0"
 [dev-dependencies]
 mockall = "0.11.0"
 
+[build-dependencies]
+cmake = "0.1.48"
+
 [[bin]]
 name = "acs"
 path = "src/main.rs"
