+++ build.rs
@@ -1,4 +1,7 @@
+extern crate cmake;
 use std::process::Command;
+use cmake::Config;
+
 fn main() {
     let output = Command::new("git")
         .args(&["rev-parse", "HEAD"])
@@ -7,4 +10,9 @@ fn main() {
 
     let git_hash = String::from_utf8(output.stdout).unwrap();
     println!("cargo:rustc-env=GIT_HASH={}", git_hash);
+
+    let dst = Config::new("libmsrtools").build();
+
+    println!("cargo:rustc-link-search=native={}", dst.display());
+    println!("cargo:rustc-link-lib=static=msrtools");
 }
