+++ msrtools/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.0)
+project(LibMSRTools C)
+add library(msrr STATIC msrr.c)
+install(TARGETS msrr DESTINATION .)
+
